Move PodStatus->TaskRunStatus code to pkg/pod
This allows us to unexport more helper functions.

Some consts are still exported, but I think those might be useful to
some callers anyway.
